Safari-611.1.10/LayoutTests/fast/performance/performance-object-gc-expected.txt

Tests that the Performance wrapper stays alive for as long as its frame.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.performance.__proto__ is Performance.prototype
PASS window.performance.customProperty is undefined.
window.performance.customProperty = 1
PASS window.performance.customProperty is 1
PASS window.performance.customProperty is 1
PASS window.performance.customProperty is 1
PASS window.performance.customProperty is 1
PASS successfullyParsed is true

TEST COMPLETE
